requirements.txt
gitpython
absl-py
